+++ example/phoenix.php
@@ -7,12 +7,20 @@ return [
     'environments' => [
         'sqlite' => [
             'adapter' => 'sqlite',
-            'dsn' => 'sqlite::memory:',
+            'dsn' => 'sqlite:' . __DIR__ . '/../testing_migrations/phoenix.sqlite',
         ],
         'mysql' => [
             'adapter' => 'mysql',
             'host' => 'localhost',
             'username' => 'root',
+            'password' => '123',
+            'db_name' => 'phoenix',
+            'charset' => 'utf8',
+        ],
+        'pgsql' => [
+            'adapter' => 'pgsql',
+            'host' => 'localhost',
+            'username' => 'postgres',
             'password' => '123',
             'db_name' => 'phoenix',
             'charset' => 'utf8',